Rename masked_register_translation.py to _masked_register_translation.py
