+++ skimage/registration/_masked_register_translation.py
